Firedrake backend: Move solver code
