+++ docs/source/conf.py
@@ -14,7 +14,6 @@ autoapi_ignore = ["*/checkpoint_schedules/__init__.py",
                   "*/fenics/parameters.py",
                   "*/firedrake/__init__.py",
                   "*/firedrake/backend.py",
-                  "*/firedrake/backend_code_generator_interface.py",
                   "*/firedrake/backend_patches.py",
                   "*/firedrake/parameters.py",
                   "*/patch.py"]
