+++ tlm_adjoint/firedrake/solve.py
@@ -1,18 +1,17 @@
-"""This module implements finite element calculations. In particular the
-:class:`.EquationSolver` class implements the solution of finite element
-variational problems.
+"""Finite element variational problem solution operations with Firedrake.
 """
 
-from .backend import adjoint, backend_DirichletBC, parameters
+from .backend import (
+    LinearSolver, adjoint, backend_DirichletBC, backend_solve, extract_args,
+    parameters)
 from ..interface import (
     check_space_type, is_var, var_axpy, var_copy, var_id, var_new,
     var_new_conjugate_dual, var_replacement, var_update_caches, var_zero)
-from .backend_code_generator_interface import assemble_linear_solver, solve
 
 from ..caches import CacheRef
 from ..equation import ZeroAssignment
 
-from .assembly import assemble, matrix_multiply
+from .assembly import _assemble_system, assemble, matrix_multiply
 from .caches import assembly_cache, is_cached, linear_solver_cache, split_form
 from .expr import (
     ExprEquation, derivative, eliminate_zeros, expr_zero, extract_coefficients,
@@ -27,8 +26,113 @@ import ufl
 
 __all__ = \
     [
-        "EquationSolver"
+        "EquationSolver",
+        "linear_solver"
     ]
+
+
+def assemble_linear_solver(A_form, b_form=None, bcs=None, *,
+                           form_compiler_parameters=None,
+                           linear_solver_parameters=None):
+    if bcs is None:
+        bcs = ()
+    elif isinstance(bcs, backend_DirichletBC):
+        bcs = (bcs,)
+    if form_compiler_parameters is None:
+        form_compiler_parameters = {}
+    if linear_solver_parameters is None:
+        linear_solver_parameters = {}
+
+    A, b = _assemble_system(
+        A_form, b_form=b_form, bcs=bcs,
+        form_compiler_parameters=form_compiler_parameters,
+        mat_type=linear_solver_parameters.get("mat_type", None))
+
+    solver = linear_solver(A, linear_solver_parameters)
+
+    return solver, A, b
+
+
+def linear_solver(A, linear_solver_parameters):
+    """Construct a :class:`firedrake.linear_solver.LinearSolver`.
+
+    :arg A: A :class:`firedrake.matrix.Matrix`.
+    :arg linear_solver_parameters: Linear solver parameters.
+    :returns: The :class:`firedrake.linear_solver.LinearSolver`.
+    """
+
+    if "tlm_adjoint" in linear_solver_parameters:
+        linear_solver_parameters = dict(linear_solver_parameters)
+        tlm_adjoint_parameters = linear_solver_parameters.pop("tlm_adjoint")
+        options_prefix = tlm_adjoint_parameters.get("options_prefix", None)
+        nullspace = tlm_adjoint_parameters.get("nullspace", None)
+        transpose_nullspace = tlm_adjoint_parameters.get("transpose_nullspace",
+                                                         None)
+        near_nullspace = tlm_adjoint_parameters.get("near_nullspace", None)
+    else:
+        options_prefix = None
+        nullspace = None
+        transpose_nullspace = None
+        near_nullspace = None
+    return LinearSolver(A, solver_parameters=linear_solver_parameters,
+                        options_prefix=options_prefix,
+                        nullspace=nullspace,
+                        transpose_nullspace=transpose_nullspace,
+                        near_nullspace=near_nullspace)
+
+
+def solve(*args, **kwargs):
+    if not isinstance(args[0], ufl.classes.Equation):
+        return backend_solve(*args, **kwargs)
+
+    eq, x, bcs, J, Jp, M, form_compiler_parameters, solver_parameters, \
+        nullspace, transpose_nullspace, near_nullspace, options_prefix = \
+        extract_args(*args, **kwargs)
+    check_space_type(x, "primal")
+    if bcs is None:
+        bcs = ()
+    elif isinstance(bcs, backend_DirichletBC):
+        bcs = (bcs,)
+    if form_compiler_parameters is None:
+        form_compiler_parameters = {}
+    if solver_parameters is None:
+        solver_parameters = {}
+
+    if "tlm_adjoint" in solver_parameters:
+        solver_parameters = dict(solver_parameters)
+        tlm_adjoint_parameters = solver_parameters.pop("tlm_adjoint")
+
+        if "options_prefix" in tlm_adjoint_parameters:
+            if options_prefix is not None:
+                raise TypeError("Cannot pass both options_prefix argument and "
+                                "solver parameter")
+            options_prefix = tlm_adjoint_parameters["options_prefix"]
+
+        if "nullspace" in tlm_adjoint_parameters:
+            if nullspace is not None:
+                raise TypeError("Cannot pass both nullspace argument and "
+                                "solver parameter")
+            nullspace = tlm_adjoint_parameters["nullspace"]
+
+        if "transpose_nullspace" in tlm_adjoint_parameters:
+            if transpose_nullspace is not None:
+                raise TypeError("Cannot pass both transpose_nullspace "
+                                "argument and solver parameter")
+            transpose_nullspace = tlm_adjoint_parameters["transpose_nullspace"]
+
+        if "near_nullspace" in tlm_adjoint_parameters:
+            if near_nullspace is not None:
+                raise TypeError("Cannot pass both near_nullspace argument and "
+                                "solver parameter")
+            near_nullspace = tlm_adjoint_parameters["near_nullspace"]
+
+    return backend_solve(eq, x, bcs, J=J, Jp=Jp, M=M,
+                         form_compiler_parameters=form_compiler_parameters,
+                         solver_parameters=solver_parameters,
+                         nullspace=nullspace,
+                         transpose_nullspace=transpose_nullspace,
+                         near_nullspace=near_nullspace,
+                         options_prefix=options_prefix)
 
 
 def homogenized_bc(bc):
@@ -80,6 +184,14 @@ def unpack_bcs(bcs, *, deps=None):
                   for g in bc_gs)
 
     return deps, bc_nodes, bc_gs
+
+
+_parameters = parameters.setdefault("tlm_adjoint", {})
+_parameters.setdefault("EquationSolver", {})
+_parameters["EquationSolver"].setdefault("enable_jacobian_caching", True)
+_parameters["EquationSolver"].setdefault("cache_rhs_assembly", True)
+_parameters["EquationSolver"].setdefault("match_quadrature", False)
+del _parameters
 
 
 class EquationSolver(ExprEquation):
